Should be comparing e values, not step numbers
Because each s has an independent value of e that the steps are counted from, s+1 can have a lower step number than s but a higher min_interp_e which makes the actual e value greater

+++ poet_src/Evolve/EccentricityExpansionCoefficients.h
@@ -59,7 +59,7 @@ namespace Evolve {
         std::vector<int> __step_num;
         std::vector<double> __max_e;
         std::vector<double> __accur;
-        std::vector<int> __order_switches;
+        std::vector<double> __order_switches;
         std::string __file_name;
         
         std::vector<double> load_coefficient(sqlite3* db,int m,int s);
